Add module for the common after_perform hook
